Continue type consistency improvements
- driver_DMAC: use bit field access instead of mask
- emon32: uint32_t -> size_t for loop counter
- periph_DS18B20: int -> uint8_t for loop counters
- pulse: uint8_t -> size_t for index parameters, add stddef.h

+++ src/emon32.c
@@ -201,7 +201,7 @@ static void cumulativeProcess(Emon32Cumulative_t    *pPkt,
  */
 static void datasetAddPulse(Emon32Dataset_t *pDst) {
   EMON32_ASSERT(pDst);
-  for (uint32_t i = 0; i < NUM_OPA; i++) {
+  for (size_t i = 0; i < NUM_OPA; i++) {
     pDst->pulseCnt[i] = pulseGetCount(i);
   }
 }
