+++ src/pulse.c
@@ -1,9 +1,11 @@
-#include "pulse.h"
+#include <stddef.h>
+
 #include "board_def.h"
 #include "driver_PORT.h"
 #include "driver_TIME.h"
 #include "emon32.h"
 #include "emon32_samd.h"
+#include "pulse.h"
 
 typedef enum PulseLvl_ { PULSE_LVL_LOW, PULSE_LVL_HIGH } PulseLvl_t;
 
@@ -12,7 +14,7 @@ static PulseCfg_t pulseCfg[NUM_OPA];
 static uint32_t   pinValue[NUM_OPA];
 static PulseLvl_t pulseLvlLast[NUM_OPA];
 
-PulseCfg_t *pulseGetCfg(const uint8_t index) {
+PulseCfg_t *pulseGetCfg(const size_t index) {
   /* If no pulse counters attached or index out of range, return 0 */
   if ((0 == NUM_OPA) || (index > (NUM_OPA - 1u))) {
     return 0;
@@ -21,7 +23,7 @@ PulseCfg_t *pulseGetCfg(const uint8_t index) {
   return &pulseCfg[index];
 }
 
-void pulseInit(const uint8_t index) {
+void pulseInit(const size_t index) {
   const uint8_t opaPUs[] = {PIN_OPA1_PU, PIN_OPA2_PU};
 
   const uint8_t pin = pulseCfg[index].pin;
@@ -42,11 +44,11 @@ void pulseInit(const uint8_t index) {
   pulseLvlLast[index] = (PulseLvl_t)pinValue[index];
 }
 
-void pulseSetCount(const uint8_t index, const uint64_t value) {
+void pulseSetCount(const size_t index, const uint64_t value) {
   pulseCount[index] = value;
 }
 
-uint64_t pulseGetCount(const uint8_t index) { return pulseCount[index]; }
+uint64_t pulseGetCount(const size_t index) { return pulseCount[index]; }
 
 void pulseUpdate(void) {
   uint32_t   mask;
